+++ python/binding/facile.c
@@ -1,12 +1,13 @@
-#include <stdlib.h>
-#include <caml/mlvalues.h>
-#include <caml/callback.h>
 #include <caml/alloc.h>
-#include <caml/memory.h>
+#include <caml/callback.h>
 #include <caml/fail.h>
-#include <stdio.h>
+#include <caml/memory.h>
+#include <caml/mlvalues.h>
 
-#include "interface.h"
+#include <stdio.h>
+#include <stdlib.h>
+
+#include "facile.h"
 
 #define CLOSURE(A)                 \
   static value *closure = NULL;    \
